requirements.txt
django
pillow
django-multiselectfield